move ticketing tests to devconnect.spec.ts
